Refactor grid click input
